+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-* Added `aioplus.atabulate`.
+* Added `aioplus.atabulate`;
+* Added `aioplus.azip`.
 
 ## [0.6.0] - 2025-08-31